requirements/dev.txt
-r common.txt

diff-cover==0.8.6
fakeredis==0.6.2
freezegun==0.3.5
pytest-cov==2.2.0
pytest-flakes==1.0.1
pytest-pep8==1.0.6
pytest-tornado==0.4.4
pytest==2.8.5
tox==2.3.1
